refactor(web): move auth support files into route group
