+++ web/app/(auth)/forgot-password/page.tsx
@@ -2,9 +2,9 @@
 import { cn } from '@langgenius/dify-ui/cn'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import * as React from 'react'
+import Header from '@/app/(auth)/_components/header'
 import ChangePasswordForm from '@/app/forgot-password/ChangePasswordForm'
 import ForgotPasswordForm from '@/app/forgot-password/ForgotPasswordForm'
-import Header from '@/app/signin/_header'
 import useDocumentTitle from '@/hooks/use-document-title'
 import { useSearchParams } from '@/next/navigation'
 import { systemFeaturesQueryOptions } from '@/service/system-features'
