+++ web/app/(auth)/reset-password/check-code/page.tsx
@@ -4,9 +4,9 @@ import { toast } from '@langgenius/dify-ui/toast'
 import { RiArrowLeftLine, RiMailSendFill } from '@remixicon/react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { createAuthSearchParams } from '@/app/(auth)/_utils/post-login-redirect'
 import Input from '@/app/components/base/input'
 import Countdown from '@/app/components/signin/countdown'
-import { createAuthSearchParams } from '@/app/signin/utils/post-login-redirect'
 import { useLocale } from '@/context/i18n'
 import { useRouter, useSearchParams } from '@/next/navigation'
 import { sendResetPasswordCode, verifyResetPasswordCode } from '@/service/common'
