+++ web/app/(auth)/reset-password/page.tsx
@@ -5,9 +5,9 @@ import { RiArrowLeftLine, RiLockPasswordLine } from '@remixicon/react'
 import { noop } from 'es-toolkit/function'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { createAuthSearchParams } from '@/app/(auth)/_utils/post-login-redirect'
 import Input from '@/app/components/base/input'
 import { COUNT_DOWN_KEY, COUNT_DOWN_TIME_MS } from '@/app/components/signin/countdown'
-import { createAuthSearchParams } from '@/app/signin/utils/post-login-redirect'
 import { emailRegex } from '@/config'
 import { useLocale } from '@/context/i18n'
 import useDocumentTitle from '@/hooks/use-document-title'
