+++ web/app/(auth)/signin/_components/normal-form.tsx
@@ -5,18 +5,18 @@ import { useSuspenseQuery } from '@tanstack/react-query'
 import * as React from 'react'
 import { useCallback, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import Split from '@/app/(auth)/_components/split'
+import Loading from '@/app/components/base/loading'
 import { IS_CE_EDITION } from '@/config'
 import Link from '@/next/link'
 import { useSearchParams } from '@/next/navigation'
 import { invitationCheck } from '@/service/common'
 import { systemFeaturesQueryOptions } from '@/service/system-features'
 import { LicenseStatus } from '@/types/feature'
-import Loading from '../components/base/loading'
-import MailAndCodeAuth from './components/mail-and-code-auth'
-import MailAndPasswordAuth from './components/mail-and-password-auth'
-import SocialAuth from './components/social-auth'
-import SSOAuth from './components/sso-auth'
-import Split from './split'
+import MailAndCodeAuth from './mail-and-code-auth'
+import MailAndPasswordAuth from './mail-and-password-auth'
+import SocialAuth from './social-auth'
+import SSOAuth from './sso-auth'
 
 const NormalForm = () => {
   const { t } = useTranslation()
