+++ web/app/(auth)/signin/check-code/page.tsx
@@ -5,10 +5,10 @@ import { toast } from '@langgenius/dify-ui/toast'
 import { RiArrowLeftLine, RiMailSendFill } from '@remixicon/react'
 import { useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { createAuthSearchParams, resolvePostLoginRedirect } from '@/app/(auth)/_utils/post-login-redirect'
 import { trackEvent } from '@/app/components/base/amplitude'
 import Input from '@/app/components/base/input'
 import Countdown from '@/app/components/signin/countdown'
-import { createAuthSearchParams, resolvePostLoginRedirect } from '@/app/signin/utils/post-login-redirect'
 
 import { useLocale } from '@/context/i18n'
 import { useRouter, useSearchParams } from '@/next/navigation'
