+++ web/app/(auth)/signin/invite-settings/page.tsx
@@ -8,9 +8,9 @@ import { useSuspenseQuery } from '@tanstack/react-query'
 import { noop } from 'es-toolkit/function'
 import { useCallback, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import { resolvePostLoginRedirect } from '@/app/(auth)/_utils/post-login-redirect'
 import Input from '@/app/components/base/input'
 import Loading from '@/app/components/base/loading'
-import { resolvePostLoginRedirect } from '@/app/signin/utils/post-login-redirect'
 import { LICENSE_LINK } from '@/constants/link'
 import { setLocaleOnClient } from '@/i18n-config'
 import { languages, LanguagesSupported } from '@/i18n-config/language'
