+++ web/app/(auth)/signin/layout.tsx
@@ -4,7 +4,7 @@ import type { ReactNode } from 'react'
 import { cn } from '@langgenius/dify-ui/cn'
 import { useSuspenseQuery } from '@tanstack/react-query'
 
-import Header from '@/app/signin/_header'
+import Header from '@/app/(auth)/_components/header'
 import useDocumentTitle from '@/hooks/use-document-title'
 import { systemFeaturesQueryOptions } from '@/service/system-features'
