+++ web/app/(auth)/signin/next/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import OneMoreStep from '@/app/signin/one-more-step'
+import OneMoreStep from '@/app/(auth)/signin/_components/one-more-step'
 
 export default function SignInNextPage() {
   return <OneMoreStep />
