+++ web/app/(auth)/signup/_components/input-mail.tsx
@@ -5,8 +5,8 @@ import { toast } from '@langgenius/dify-ui/toast'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { useCallback, useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import Split from '@/app/(auth)/_components/split'
 import Input from '@/app/components/base/input'
-import Split from '@/app/signin/split'
 import { emailRegex } from '@/config'
 import { useLocale } from '@/context/i18n'
 import Link from '@/next/link'
