+++ web/app/(auth)/signup/page.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
-import { createAuthSearchParams } from '@/app/signin/utils/post-login-redirect'
-import MailForm from '@/app/signup/components/input-mail'
+import { createAuthSearchParams } from '@/app/(auth)/_utils/post-login-redirect'
+import MailForm from '@/app/(auth)/signup/_components/input-mail'
 import { useRouter, useSearchParams } from '@/next/navigation'
 
 const Signup = () => {
