+++ web/app/activate/activateForm.tsx
@@ -3,9 +3,9 @@ import { Button } from '@langgenius/dify-ui/button'
 import { cn } from '@langgenius/dify-ui/cn'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
-import Loading from '@/app/components/base/loading'
+import { createAuthSearchParams } from '@/app/(auth)/_utils/post-login-redirect'
 
-import { createAuthSearchParams } from '@/app/signin/utils/post-login-redirect'
+import Loading from '@/app/components/base/loading'
 import useDocumentTitle from '@/hooks/use-document-title'
 import { useRouter, useSearchParams } from '@/next/navigation'
 import { useInvitationCheck } from '@/service/use-common'
