--- web/app/signin/page.module.css
@@ -1,11 +1,0 @@
-.githubIcon {
-  background: center/contain url('./assets/github.svg') no-repeat;
-}
-
-html[data-theme="dark"] .githubIcon {
-  background: center/contain url('./assets/github-dark.svg') no-repeat;
-}
-
-.googleIcon {
-  background: center/contain url('./assets/google.svg') no-repeat;
-}
